Add smetamath (in Rust) as 3rd Travis metamath validator
This modifies .travis.yml (the configuration file for running Travis)
so that changes will also be checked by smetamath.
The smetamath tool (aka smetamath-rs and smm3) is a high-speed
metamath verifier written in Rust by Stefan O'Rear that can use multiple cores.

This means that set.mm is now checked by three
independently-implemented metamath verifiers in three different languages:
metamath (in C), mmj2 (in Java), and smetamath (in Rust).
These are also all written by different people.
This use of diverse verifiers greatly descreases the likelihood
that a code defect in a verifier might result in
incorrect validation of metamath proofs.

My thanks to Stefan O'Rear, both for the tool and for various tweaks
that make this work.