Change reference from df-sgn to df-psgn
Change reference from df-sgn to df-psgn in David A. Wheeler's mathbox.

Gérard Lang-Marconnet stated that Stefan's definition of the
sign of a permutation, df-psgn, was superior to the definition
given in df-sgn.

Mario Carneiro agreed and said, "I thought I deleted that ages ago.
It can't handle permutations of infinite sets, and it may be
incorrect beyond that. I'm supporting Stefan's definition 100%."

However, David A. Wheeler's mathbox definition df-sgnf referenced df-sgn
(merely as a constrast).  This commit removes that reference to df-sgn,
which should pave the way for removing df-sgn entirely.