Move Aristotelian syllogisms into main body of set.mm
Move Aristotelian syllogisms into the main body from the mathbox of
David A. Wheeler.  This shows that the selected axioms are enough to
prove these syllogisms, as well as provide useful theorems.
This moves the proof alsyl contributed by
Andrew Salmon (PM theorem *10.3) into the main body.
It also includes updates to the section introduction text.