Move > and >_ from David A. Wheeler's mathbox to main body; use discouraged
Per David A. Wheeler's email on Sun, 01 Jan 2017 14:42:45 -0500 (EST),
move greater-than ` > ` and greater-than-or-equal ` >_ ` into the
main body with all relevant uses marked as "(New usage is discouraged.)".

"Greater-than" and "greater-than-or-equal" are widely used.  Even in
set.mm the term "greater" is used 185 times when you ignore this my
greater-than section.  I'd like set.mm to include formal definitions
of all significant mathematical terms (since I view its ultimate
goal as "formally proving all mathematics using typical ZFC axiomatic
conventions").  In that light, omitting greater-than is a problem.

The old justification for *not* moving it up is reasonable enough:
Every use creates additional work, so the convention is to always use <
and <_. Obviously, omitting ">" and ">_" from the main body ensures that
"<" and "<_" are always used instead.  But these are now marked
with the newer marking "(New usage is discouraged.)".
With that marking, I believe that problem goes away. This
marking will discourage any use of > and >_ in new theorems. In addition,
even if that is ignored, the set.mm integration tests now *specifically*
check to make sure that all changes to the "discouraged" list are
noticed. That means that > and >_ can be moved into the main body
*without* any risk that they will be unintentionally used.

The examples were moved into the "examples" section.