In .travis.yml, modify most wget's to add "-N" option
By default wget will rename files if there's a file by that name already.
This is often counterproductive, since we often simply want to update
the old file with the new one... this default means we will often
accidentally use the old version.  In addition, this behavior forces
downloading of files even if we already have the same file
(as determined via timestamps).

This change means that wget will check timestamps of the downloaded
files, only download the rest of the file if the timestamps are
different, and replace the old file if they are different.

This situation doesn't happe in the current travis setup because we
always start with a clean slate.  However, this change means that people
can easily cut-and-paste the travis script to use on their own system,
and it means that we can usefully enable caching in Travis.