Wallis' formula for pi and Stirling's formula for n!
The main results in this commit are the Wallis' product for Pi and
Stirling's approximation formula for n factorial. Stirling's formula is
#90 in the 100 theorems list.
I've proven two versions of Wallis' product and two versions of
Stirling's formula: stirling proves the formula w.r.t. the topology on
complex numbers,
and stirlingr proves the formula w.r.t. the standard topology on the
reals.