Prefer "addcl" over both "ax-addcl" and "axaddcl"
Expressly mark "ax-addcl" with "(New usage is discouraged.)",
and make it clear that by convention we normally use "addcl"
instead of both "ax-addcl" and "axaddcl".
They all express the same statement, but "addcl" follows
our naming conventions for closures, so we normally want to use
that one.

In short, we first prove "axaddcl".  We then re-assert it as
axiom "ax-addcl".  We do that because this follows our
axiom and proven axiom naming conventions.  But this is also
a closure, so we restate it as "addcl" using the closure naming
conventions, add from then on prefer that name.
By doing it this way, we follow a standard set of conventions
everywhere that should make the whole thing easier to follow.
Since we expressly mark "ax-addcl" with "(New usage is discouraged.)",
tools can prefer the "normally correct" theorem by our conventions.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>