Discourage new use of axlttrn ; use lttr instead
axlttrn and its alias lttr show that ordering on reals is transitive.

For naming consistency, as well as simply using a shorter name,
we'll prefer "lttr" over "axlttrn".  We'll enforce this convention
by marking axlttrn so new usage is discouraged.

The main spine for proving that ordering on reals is transitive is:
axpre-lttrn (new usage already discouraged)
ax-pre-lttrn $a - uses <RR, which is only for non-extended reals
axlttrn (add new usage is discouraged)
lttr (the preferred one)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>