Note in "conventions" how to add new references
We want to encourage people to add appropriate citations.
To help, modify the "conventions" section to more
clearly document how to add a new reference
(by adding it to mmset.html).  Also, make it clear that
the file mmbiblio.html is automatically rebuilt; new users
might not realize that it is automatically generated.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>