Discourage new use of <RR, ax-addf, ax-mulf
This discourages use of the <RR relation - proofs are instead
encouraged to use the more common < relation instead.
We do this by discouraging new use of a few axioms with <RR
that weren't already discouraged, and documenting what is
recommended instead.

This also discourages future use of deprecated
axioms ax-addf and ax-mulf (see their comments for more)>

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>