Add table showing common label abbreviations
Add a table showing the common abbreviations used in labels.
The table notes if the abbreviation is based on syntax, and
includes examples of the abbreviation.  The examples were
selected so that users can often decipher an entire label
using just the information in the "conventions" section,
which should make the conventions much easier to understand.

This puts the naming conventions in *one* place within set.mm
so that if the naming convention changes, the names and the
naming convention documentation can be updated all at once.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>