Rename "df-plus" to "df-add"
Rename "df-plus" to "df-add" so the name is consistent with labels
that refer to it.  Label fragments already consistently refer to
"+" as "add", e.g., "addcl".  Even the comment refers to the
operation as "addition".

Note that this does *not* change "df-plusg",
because relevant labels are already consistent with it
(they refer to "plusg").

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>