Add info on biconditional and prop calc names to "conventions"
Add information to "conventions" section describing the rationale
for how biconditional is handled, and the names related to
propositional calculus.  This is based on a mailing list discussion
titled "True antecedent".  The text here is from the email
of David A. Wheeler dated "Sun, 20 Aug 2017 23:09:06 -0400 (EDT)".
That text was in turn developed by several people.
A lot of this text was originally written by Mario Carneiro.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>