Move conventions and natded down, to just before htmldefs
The "conventions" and "natded" sections create very long sections
in the detailed table of contents.  This is distracting if they
are near the beginning (as they were).  So, move these to near the
end, just before the htmldefs.  That way they aren't distracting.

Note: "git diff" may show more changes than actually occurred in this
commit.  Use "git diff --minimal" to show the difference.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>