Remove/move classical theorems.
The general rule here is that if a theorem's proof depends on ax-3
and the theorem isn't used or mentioned elsewhere in the file, we
delete it (especially if it doesn't have an intuitionistic proof
or it doesn't seem important to try to find one for that particular
theorem). If it is mentioned, but not used, it should be moved to
the end of the file where we plan on eventually moving ax-3.

- Move 4cases, ax11eq, and ax11el to the end of the file
- Remove 4casesdan, consensus, and pm4.42
- Remove elimh. Since it is tied to the rest of the weak deduction
theorem machinery, also remove dedlem0a, dedlem0b, dedlema, dedlemb,
dedt, and con3th. Although some of these are valid intuitionistically,
there doesn't seem to be much point in having just part of this
group of theorems.
- Revise discussion of ax-11 to reflect even more of the theorems
mentioned there relying on ax-3.
- Write the file with /rewrap (for consistent line length).