Shorten proof of sbco2iv.
Although I wrote the sbco2iv proof, I'm still trying to figure out
just what makes it tick and what can be used elsewhere. Here's a
good first step to making it more understandable: cut out a
significant part of the proof which was based on moving around
E. z when the expression being quantified doesn't have z as a free
variable. The use of ax11ev instead of ax-i11e is for convenience;
it would be almost as easy (although a few more steps) to make use
of the ph -> A. z ph hypothesis instead.

Goes from 756 characters (compressed) to 645.