Be clearer about loolin and loolinALT.
In the context of set.mm, I suppose we might as well keep both but we
should make it clear that the statment of both theorems is the same
and it is just the proof which is different. In the case of iset.mm,
there isn't much point in having a non-intuitionistic proof.