Remove imordi in favor of pm4.78.
In set.mm, this just means removing the redundant imordi and making
the description of pm4.78 slightly more verbose. (The other set.mm
changes are just /rewrap).

In iset.mm, rename imordir to pm4.78i and have the descriptions of
it and pm4.78 refer to each other. While we're at it, move pm4.78
to the classical logic section since the current proof is classical
and it isn't currently used. Oh, and pm4.78i was wrongly in the
classical section; move that one earlier.