Change "dfnot" to be intuitionistic (do not use "ja")
This commit resolves the comment from Jim Kingdon on 23-Feb-2018:
"Found the `ax-3` usages. The proof of `dfnot` is not an intuitionistic
one (it uses `ja`). Replacing that with the `dfneg` proof should take
care of both of them (the proof of `inegd` looks fine except for the
`dfneg` problem)."

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>