Add some truth table entries and related theorems to iset.mm (#337)
* Add some truth table entries and related theorems to iset.mm

Most of these proofs are from set.mm; credit is retained (unmodified)
in those cases.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix date at top of iset.mm

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Use "dfnot" (set.mm name) and never "dfneg"

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Apply set.mm renames to iset.mm (notfal->fal, notfalbitru->notfal)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Change "dfnot" to be intuitionistic (do not use "ja")

This commit resolves the comment from Jim Kingdon on 23-Feb-2018:
"Found the `ax-3` usages. The proof of `dfnot` is not an intuitionistic
one (it uses `ja`). Replacing that with the `dfneg` proof should take
care of both of them (the proof of `inegd` looks fine except for the
`dfneg` problem)."

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update iset.mm date at top

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>