Add sbcomv and sbcocom to iset.mm.
sbcomv is a version of sbcom with another distinct variable constraint,
but with an intuitionistic proof.

For current purposes we can think of sbcocom as a lemma for sbcomv (given
sbco3v), but it also could used in the other direction (to give us sbco3
if we were to get sbcom for example).