Remove stray classical logic comments in iset.mm
There are some comments which refer to the logic as classical logic
(which it is not), or make statements about interpretation of
connectives which do not apply to intuitionistic logic. It seems
better to just remove those comments than to try to get into
intuitionistic logic equivalents (which are more complicated
and not necessary to follow the proofs here).