iset.mm: remove ax-3-using explorations of axioms
That is, remove the "Predicate logic axioms (classical)" section
at the end of the file. This contains proofs inherited from
set.mm concerning predicate logic axioms (which axioms are
redundant, which theorems depend on which axioms, etc).

This section has gotten little attention since being copied
from set.mm and also never was based on
intuitionistic logic. Future investigations of
the axioms can just as easily refer to set.mm (which also has
additional insights, such as the proof of ax-10 from other
axioms, which were not part of the initial fork which led
to iset.mm).

Remove various comments which refer to this section. Many of
them already said "in classical logic, . . ." and others
maybe should have had that disclaimer but lacked it.