Remove neor from iset.mm.
Some of the usages, like the theorem `prime`, could use a version of
neor for decidable equality (because equality of natural numbers is
decidable).  Others, like fimaxre, could not (because equality of
real numbers is not, in general, decidable).  In any event, we
can revisit this topic later, because none of those theorems are
coming up soon.