Fix forward reference in iset.mm
There are a few references in comments to theorems that we
haven't yet brought over from set.mm. It isn't clear there
is a great way to handle this situation, but leaving in
the comment but just referring to the theorem name as text,
rather than a cross reference, might be as good as anything
else.