Intuitionize mosub in iset.mm
The obvious proof in terms of sbmo seems to require an additional
A e. _V hypothesis so add the hypothesis and rename mosub to mosubex
because it no longer is the same. Although the theorem makes
sense when A is a proper class (because y = A will never be true
then), whether there is an intuitionistic proof is a different
question.

Since mosub is not used in set.mm, there isn't an issue of whether
the proofs that use it can supply the additional hypothesis.