Add ru to iset.mm
The theorem and proof are verbatim copies from set.mm.

The comment is partly copied but significantly trimmed, partly to
avoid referring to theorems which don't exist in iset.mm (yet at
least) but more importantly because parts of it might not apply,
or not apply quite as stated, in intuitionistic logic (for
example, New Foundations or Neumann-Bernays-Goedel don't
necessarily have the same relationship with IZF as ZF, or see the
discussion of large sets in Crosilla section 5.2 with respect
to the mention of large sets and New Foundations).

The citation to Crosilla is formatted slightly oddly to avoid
a warning from VERIFY MARKUP */F. It seems to sometimes
insist on a page number, but I'm not sure what its exact rules
are or how to apply them to a web work which, although it has
a whole system of citing specific versions, does not have page
numbers (at least in the usual way of reading it).