Minimize out wsbc in iset.mm
Apply minimize_with * to all proofs which use wsbc. Because they
are all (currently) able to use wsb instead, that means there
are no longer any usages of wsbc. That means we'll be able to
harmonize with set.mm (in which substituting a class or
substituting a variable have different syntaxes, not overloaded).