Remove dfsb2 and dfsb3 from iset.mm
There doesn't seem to be any need to say anything about dfsb3
because the point is to use primitive connectives (and what it
considers to be primitive is not the same as iset.mm).

As for dfsb2, it would probably not be especially easy to
come up with a version which uses DECID or other conditions,
and if we did, the extra conditions might be clutter as much
as something which shows the contrast between classical and
intuitionistic logic. So putting the definition in a comment
and not trying to prove anything about it may be the best
option.

Reduces "show usage ax-3/recursive" count by two.