Remove set operation examples from iset.mm
This is ex-dif , ex-in , and ex-un together with the statement of
those examples (which depends on syntax we don't yet have).

It will be some time before we have the right machinery (like,
numbers of any kind) to make those examples work, and there
isn't really an obvious way to do them with what we have at
hand right now.