Remove qexmid from iset.mm
For theorems which we don't have an intuitionistic proof for,
and which aren't used in set.mm, and which don't have a clear
way to add a decidability hypothesis, just removing them from
iset.mm seems like a good plan. They still exist in set.mm
so we can always see them there.