Add axioms and theorems to mmil.html
Add set theory subsection to the axioms section. Right now it just
has extensionality, but we'll add the other IZF axioms as we get
to them.

Also add Theorem Sampler.  It is based on the theorem sampler
from mmset.html but adapted for intuitionistic logic.  We only
can get a bit into set theory but this too can be expanded as
we have more of set theory.