Add alternative `axiom` in formal system developed by Heyting
I'm new one publisher and enthusiast of Metamath project and I've chosen this missed theorem to add as a training example.
I'd like to contribute more but I believe that due to lack of experience and language my changes and naming of those changes can be the complete disaster.
That's why your review would be especially appreciated, because any feedback will be useful!
Also I've only russian edition of Heyting introduction to Intuionistic so page link could be misleading. If you can help me with this I'd really appreciate that!
Thanks!