Restore full mo2icl to iset.mm
Initially iset.mm had a weakened version ( mo2iclex ) which
required that A exist. Restore the full theorem as found in
set.mm but with an intuitionistic proof. The proof somewhat
resembles the proof from set.mm but case elimination had to
be replaced with moabs and some other logic.