Remove disjiun and disjiunOLD from iset.mm
The current proof of disjiun relies on rmo2 which we don't have
in iset.mm.

They are not used enough in set.mm that we need to address
this now.  disjiun is used by disjxiun but disjxiun is not used
anywhere, and disjiun has a few usages much later in the
file.