Add axioms of collection and separation to iset.mm
Add ax-coll , the Axiom of Collection as stated by Crosilla.

Add repizf which is not an axiom in IZF but which is an
easy consequence of ax-coll .

Add zfrep6 which is very similar to ax-coll but which
is worth stating as it exists in set.mm.

Add ax-sep - as stated in set.mm which is different in minor
ways from Crosilla's formulation.  Unlike in set.mm this is
an axiom not a theorem because we have ax-coll/zfrep6 instead
of set.mm's ax-rep.