Intuitionize iinexg in iset.mm
Rename to iinexgm and change non-empty to inhabited. Also,
the proof, although basically the same in structure, needs
changes throughout for intuitionistic logic. Specifically,
r19.35-1 in place of r19.35 , r19.2m in place of r19.2z ,
abid in place of abn0 , and inteximm in place of intex
(which also means we need a variable change).