Intuitionize snexALT in iset.mm
The best I've been able to do so far is separate theorems for
the A e. _V and -. A e. _V cases, named snexg and snexprc .
Given how many places snex is used, it would be nice if we
could prove the original snex . However, the case elimination
seems a bit fundamental, so we'll see whether snex is possible.