Remove ax16b and intuitionize eunex in iset.mm
ax16b makes statements about independence which might not apply
to iset.mm, so we are better off leaving it to set.mm.

Prove eunex with the aid of a hypothesis which is a version of
dtru (which hopefully we'll be able to prove when someone
figures out how). Because it has this hypothesis, rename to
eunexdtru .