Tweak class history section of mmset.thml
Tweak the history paragraph of the class section in mmset.html
(the HTML text explaining set.mm).  In particular, this makes it
clear that the "virtual classes" used in set.mm are *NOT* the same as
von Neumann-Bernays-Gödel (NBG), and points to [Quine 1963] so
that interested readers will know where to go to get a detailed
examanation of the differences between this "virtual class" approach
and some other approaches.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>