Remove opthwiener from iset.mm
The current proof relies on snex and prex . Like opth , opthwiener
does not have C e. _V or D e. _V as hypotheses, so moving to
snexg , prexg , etc, would be somewhat involved. Since opthwiener
is not used in set.mm, there is no particular reason this has to
be tackled now, so remove it from iset.mm until someone takes
this up.