Discourage depending on the Kuratowski definition for ordered pairs
We use the Kuratowski definition for ordered pairs, but
there are many other ways to define ordered pairs.
For example, the Kuratowski definition is very inconvenient in
New Foundations theory because it is not type-level, so
the Definition from [Rosser] p. 281 is often used instead in that context.

This commit uses "(New usage is discouraged.)" markings to discourage
depending on the particular definition/construction of ordered pairs.

This way, most theorems that use ordered pairs will not depend
on this particular Kuratowski construction, but will instead only
depend on more abstract properties of ordered pairs.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>