Remove well-foundedness from iset.mm
Remove the well-foundedness and well-ordered predicates and
all definitions and theorems which depend on them. It is
possible they could come back (with non-empty changed to
inhabited, presumably), but a number of proofs would need to
change and I'm not sure these concepts are relevant.

Rename the "Founded and well-ordering relations" section to
"Set-like relations" as that is all that is left. The tz7.2
theorem didn't seem to be about well-foundedness but it
does have a well-foundedness condition; set.mm does not
remove that condition until trelpss .