Add algebra helpers and "i2lines" to demo it
This commit adds to David A. Wheeler's mathbox a collection of
"algebra helpers" that I suspect will make it easier to do
a number of simple algebra operations on equations.
It also includes "i2lines", which demonstrates how to use them.

This is an experiment in a different approach, but since all theorems are
proven and no new definitions are added, everything is of course proven.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>