Improve Algebra helpers in David A. Wheeler's mathbox per Benoit's comments
Remove some unnecessary $e hypotheses in the Algebra helpers
and modify the corresponding i2lines example.
This is in reponse to some helpful comments by Benoit (@benjub).
This is experimental work in David A. Wheeler's mathbox to
develop a higher-level and easier way to express simple algebra
manipulations of equations.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>