Add "Ordinals (continued)" section to iset.mm
This is just a small part of the corresponding section from
set.mm. In fact, only three theorems survive (pretty much
everything else is either patently non-intuitionistic, or
depends on things which we haven't been able to prove yet).

The three are sucexb , sucexg , and sucex .

For those three, this is an unchanged copy from set.mm.
Follow up commits will be needed to get it to verify.