Update surprises section in David A. Wheeler's mathbox
Update the "surprises" section.  For example,
"empty-surprises" didn't have anything directly to do with the empty set.
so this renames it to alimp-surprise and creates a new "empty-surprises"
that *does* have a direct relationship with the empty set.
This is partly in response to helpful feedback from
Norm Megill and Reg Meeson (thank you!).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>