Remove cases from induction from iset.mm
This is for the forms of transfinite induction which have
separate cases for successor and limit. Those work in
classical logic, but in intuitionistic logic we only get
the ones which rely on an induction step of "the property
holds for all ordinals less than this one implies that it
holds for this one".

The removed theorems are tfinds , tfindsg , tfindsg2 ,
tfindes , tfindes , and tfinds3