Document wff symbol variations in conventions
Document in "conventions" that
we try to follow mathematical conventions, but in many cases
different texts use different conventions.
In those cases we pick some reasonably common convention and stick to it.
Provide specific details about the variables used to represent
arbitrary well-formed formulas (wffs), showing that there is a wide
variety in practice.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>