Add ordsucim to iset.mm
This is the forward direction of ordsuc and is fairly similar to
suceloni . It is possible that a proof similar to the set.mm
proof of suceloni (based on trssord ) would work, but this proof
is of similar length and perhaps more readable. A downside is
that it does rely on df-iord directly, so if we picked a different
definition of Ord we'd need to change it.