Change limit ordinal definition in iset.mm
We'd expect at least the usual change of non-empty to inhabited.
However, we take it one step further, and adopt a definition
( dflim2 ) that to be a limit ordinal an ordinal must contain the
empty set, not just any element. Being inhabited might imply
containing the empty set, but I was not able to find a proof so
there isn't any particular problem in using the "contain the
empty set" definition.

Adjust theorems which had been using df-lim to use dflim2 instead
( limeq , nlim0 , limord , limuni ). Add 0ellim and prove in
terms of dflim2 .