Add Peano postulates to iset.mm
This is peano1 , peano2 , peano3 , peano4 , and peano5 .

In the case of peano1 , peano2 and peano5 , prove from dfom3
and make
no attempt to avoid the axiom of infinity. This might reflect
what is possible in IZF in general, but it certainly reflects
what is possible/easy given the current state of ordinals in
iset.mm.

In the case of peano3 , the proof is similar to set.mm but
only needs to be changed to use nsuceq0g in place of nsuceq0 .

In the case of peano4 , prove in terms of suc11g . We'll
presumably be proving nnon soon but it might be easier to
do it after the Peano axioms.