Use IZF and set induction terminology
This is in iset.mm.

There are a few places where we had been saying ZF where IZF makes
more sense.

Also, the term "set induction" seems to be at least as common as
"e.-induction" (the more I read the literature) and is less
awkward especially in contexts where we might be using ASCII
for whatever reason. Rename ax-elind to ax-setind and elindel
to setindel .