Use set induction more in iset.mm
This expands the section "Introduce the Axiom of Set Induction"
with most of the theorems from the section "Introduce the Axiom
of Regularity" in set.mm

For preleq , rearrange the propositional logic.

For opthreg , change prex to prexg but otherwise keep the
proof the same.

suc11g is based on the set.mm suc11reg but only considers the
case where both A and B exist. The propositional logic is
also rearranged but the basic concept (of applying en2lp )
is unchanged.

suc11 is stated as in set.mm but depends on the axiom of
set induction rather than by defining ordinals to be
well-founded.

Omit dford2 - ordinal trichotomy is equivalent to the law of
the excluded middle per ordtriexmid .