Update Axiom of Set Induction references
This is two places in iset.mm where we had been speaking of the
Axiom of Regularity but mean the Axiom of Set Induction.
We continue (intentionally) to use "reg" in theorem names to
refer to a theorem which relies on the Axiom of Set Induction,
because it has similar consequences to the classical Axiom of
Regularity. But when we are talking about the axiom, we'll
call it the Axiom of Set Induction.