Add natural number theorems to iset.mm
This is omelon2 , omelon , nnon , nnoni , nnord , omsson ,
peano2b , and nnsuc .

We now have pretty much everything from the "The natural numbers
(i.e. finite ordinals)" section in set.mm except for limit
ordinals - elom , limomss , nnlim , omssnlim , limom , and ssnlim .
We haven't done much with limit ordinals and it is not clear
they are as important in IZF as in ZF. But if we do want to add
them back later, they will need new proofs.