Remove Fr , We , and if from iset.mm
We don't yet have any of these in iset.mm and there are various
theorems in the Relations section which depend on them. So
we should get rid of those theorems unless/until we add back
Fr , We , or if. The removed theorems are: frinxp , weinxp ,
frsn , wesn , dffr3 , somin1 , somincom , somin2 , soltmin ,
and xpima .