Modify set.mm to note stable/testable/decidable
In intuitionistic logic only some wffs are stable,
testable, and/or decidable (decidable means the wff is both
stable and decidable).  In classical logic, all wffs are
stable, testable, and decidable.

This commit modifies set.mm where expressions equivalent to
stable, testable, and decidable are presented so that
their proofs mention these terms.  There was no direct proof about
testability, so this commit adds a proof that all wffs are
testable as part of DAW's mathbox.

This commit provides a clearer link (and constrast) between
set.mm and iset.mm.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>