Intuitionize posn and sosn in iset.mm
In both cases add a A e. _V condition and therefore rename to
posng and sosng respectively.

As for the proofs, posng is just taken from the A e. _V case in
set.mm. sosng is different from sosn due to the difference
between df-so and df-iso but also ends up being a degenerate
case and therefore holds.