Intuitionize csbxp in iset.mm
Remove csbxp (depends, both explicitly and implicitly, on case
elimination on A e. _V).

Rename csbxpgOLD to csbxpg and remove OLD from some of the proofs
which it depends on. All of these are theorems with A e. _V
antecedents and all of them exist in iset.mm (and are expected
to stay since we don't have case elimination).