Add new scripts/unjunk.py command
Add new script named scripts/unjunk.py, which
replaces img references with `...` references.
It is a filter, that is, it reads from stdin and sends to stdout.

For example, it replaces this:
<IMG SRC='_ltbbr.gif' WIDTH=20 HEIGHT=19 ALT=' &lt;RR' TITLE='&lt;RR'>
with this:
` <RR `

The intent is to help make the mm*.html files easier to edit and update,
by making it possible to reuse the metamath typesetting machinery.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>