Add mmcomplex.raw.html, a pre-formatted version of mmcomplex.html
This adds mmcomplex.raw.html, a version of mmcomplex.html but
uses metamath `...` typesetting commands instead of
having baked-in IMG references.

This is a first cut at trying to transition these files to
something that is easier to edit and directly supports the
Unicode typesetting.

This file created by running:

> python scripts/unjunk.py < mmcomplex.html > mmcomplex.raw.html

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>