Fix mmcomplex.html and mmcomplex.raw.html
This fixes mmcomplex.html and its corresponding parts in
mmcomplex.raw.html.  It makes these changes:

* (. becomes C.
* (_ becomes C_
* P~ becomes ~P
* seq1 becomes seq
* df-seq1 becomes df-seq

At the moment, "mmcomplex.html" is the master version, so I edited it
and then generated mmcomplex.raw.html using:

> python scripts/unjunk.py < mmcomplex.html > mmcomplex.raw.html

In the long term I hope to make mmcomplex.raw.html the master
version, which then generates mmcomplex.html, but that potential
result is for another time.

The discussion about df-seq may be obsolete, but currently
I'm just trying to fix the formatting.  Updating the actual
discussion there is for another time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>