Generalize r19.3rm in iset.mm
The change is that the variable in the first half and in the
second half are not the same. Since there is no distinct
variable constraint between the two, the generalized version
can be used whereever the old one had been.

Do this by renaming r19.3rm to r19.3rmOLD and proving the
generalized theorem as r19.3rm .