Intuitionize iotaex in iset.mm
Add E! x ph antecedent and rename to euiotaex . The set.mm proof
relies on case elimination, so it looks like we might not be
able to be as clever about what happens when E! x ph does not
hold. And perhaps fvex will get a Fun F antecedent or something
like that.