Intuitionize funimaexg in iset.mm
This is the first time we do something non-trivial with
replacement and bringing the gap between the zfrep6 style of
replacement and the axrep4 style involves quite a bit of
legwork (including my favorite theorem of the hour,
mopick ).

Little or none of this is about intutionistic vs classical
logic; it is more about different ways of stating replacement
and in particular the usual formulation in IZF versus the
replacement axiom in set.mm.